Removed disabling of LootFrame mover during initialize
Disabling it because "Loot under Mouse" is enabled is fine, but there is
no code in place to enable the mover again if the user disables this
Blizzard setting. It would require a reload in order for the mover to
show up then. Simply removing the disabling is the easier choice here.